+++ lib/any_cache.rb
@@ -10,6 +10,7 @@ require 'concurrent/atomic/reentrant_read_write_lock'
 class AnyCache
   require_relative 'any_cache/version'
   require_relative 'any_cache/error'
+  require_relative 'any_cache/dumper'
   require_relative 'any_cache/drivers'
   require_relative 'any_cache/adapters'
   require_relative 'any_cache/logging'