+++ lib/any_cache/adapters/active_support_mem_cache_store.rb
@@ -51,25 +51,25 @@ module AnyCache::Adapters
     def_delegators :driver, :delete, :clear
 
     # @param key [String]
-    # @param options [Hash]
+    # @option raw [Boolean]
     # @return [Object]
     #
     # @api private
     # @since 0.2.0
     def read(key, **options)
-      raw = options.fetch(:raw, true)
+      raw = options.fetch(:raw, false)
 
       driver.read(key, raw: raw)
     end
 
     # @param keys [Array<String>]
-    # @param options [Hash]
+    # @option raw [Boolean]
     # @return [Hash]
     #
     # @api private
     # @since 0.3.0
     def read_multi(*keys, **options)
-      raw = options.fetch(:raw, true)
+      raw = options.fetch(:raw, false)
 
       driver.read_multi(*keys, raw: raw).tap do |entries|
         entries.merge!(Hash[(keys - entries.keys).zip(READ_MULTI_EMPTY_KEYS_SET)])
@@ -79,25 +79,26 @@ module AnyCache::Adapters
     # @param key [String]
     # @param value [Object]
     # @option expires_in [NilClass, Integer] Time in seconds
+    # @option raw [Boolean]
     # @return [void]
     #
     # @api private
     # @sicne 0.2.0
     def write(key, value, **options)
       expires_in = options.fetch(:expires_in, NO_EXPIRATION_TTL)
-      raw = options.fetch(:raw, true)
+      raw = options.fetch(:raw, false)
 
       driver.write(key, value, expires_in: expires_in, raw: raw)
     end
 
     # @param entries [Hash]
-    # @param options [Hash]
+    # @option raw [Boolean]
     # @return [void]
     #
     # @api private
     # @since 0.3.0
     def write_multi(entries, **options)
-      raw = options.fetch(:raw, true)
+      raw = options.fetch(:raw, false)
 
       driver.write_multi(entries, expires_in: NO_EXPIRATION_TTL, raw: raw)
     end
@@ -105,6 +106,7 @@ module AnyCache::Adapters
     # @param key [String]
     # @option expires_in [Integer]
     # @option force [Boolean, Proc]
+    # @option raw [Boolean]
     # @return [Object]
     #
     # @api private
@@ -113,8 +115,9 @@ module AnyCache::Adapters
       force_rewrite = options.fetch(:force, false)
       force_rewrite = force_rewrite.call(key) if force_rewrite.respond_to?(:call)
       expires_in    = options.fetch(:expires_in, NO_EXPIRATION_TTL)
+      raw           = options.fetch(:raw, false)
 
-      driver.fetch(key, force: force_rewrite, expires_in: expires_in, &fallback)
+      driver.fetch(key, force: force_rewrite, expires_in: expires_in, raw: raw, &fallback)
     end
 
     # @param keys [Array<String>]
@@ -151,7 +154,7 @@ module AnyCache::Adapters
       expires_in = options.fetch(:expires_in, NO_EXPIRATION_TTL)
 
       unless exist?(key)
-        write(key, amount, expires_in: expires_in) && amount
+        write(key, amount, expires_in: expires_in, raw: true) && amount
       else
         driver.increment(key, amount).tap do
           expire(key, expires_in: expires_in) if expires_in
@@ -174,7 +177,8 @@ module AnyCache::Adapters
         #   - non-raw values;
         #   - values lower than zero;
         #   - empty entries;
-        write(key, INITIAL_DECREMNETED_VALUE, expires_in: expires_in) && INITIAL_DECREMNETED_VALUE
+        write(key, INITIAL_DECREMNETED_VALUE, expires_in: expires_in, raw: true)
+        INITIAL_DECREMNETED_VALUE
       else
         driver.decrement(key, amount).tap do
           expire(key, expires_in: expires_in) if expires_in
@@ -189,9 +193,9 @@ module AnyCache::Adapters
     # @api private
     # @since 0.2.0
     def expire(key, expires_in: DEAD_TTL)
-      read(key).tap do |value|
+      read(key, raw: true).tap do |value|
         is_alive = expires_in ? expires_in.positive? : false
-        is_alive ? write(key, value, expires_in: expires_in) : delete(key)
+        is_alive ? write(key, value, expires_in: expires_in, raw: true) : delete(key)
       end
     end