+++ lib/any_cache/adapters/basic.rb
@@ -6,6 +6,8 @@ module AnyCache::Adapters
   class Basic
     # @since 0.1.0
     extend Forwardable
+    # @since 0.4.0
+    include AnyCache::Dumper::InterfaceAccessMixin
 
     class << self
       # @param driver [Object]