+++ lib/any_cache/adapters/redis_store.rb
@@ -18,15 +18,16 @@ module AnyCache::Adapters
     def_delegators :driver, :mset
 
     # @param key [String]
-    # @param options [Hash]
+    # @option raw [Boolean]
     # @return [Object]
     #
     # @api private
     # @since 0.1.0
     def read(key, **options)
-      raw = options.fetch(:raw, true)
+      raw = options.fetch(:raw, false)
+      value = get(key, raw: true)
 
-      get(key, raw: raw)
+      raw ? value : detransform_value(value)
     end
 
     # @param keys [Array<String>]
@@ -43,24 +44,29 @@ module AnyCache::Adapters
     # @param key [String]
     # @param value [Object]
     # @option expires_in [NilClass, Integer] Time in seconds
+    # @option raw [Boolean]
     # @return [void]
     #
     # @api private
     # @since 0.1.0
     def write(key, value, **options)
       expires_in = options.fetch(:expires_in, NO_EXPIRATION_TTL)
-      raw = options.fetch(:raw, true)
+      raw = options.fetch(:raw, false)
+      value = transform_value(value) unless raw
 
-      expires_in ? setex(key, expires_in, value, raw: raw) : set(key, value, raw: raw)
+      expires_in ? setex(key, expires_in, value, raw: true) : set(key, value, raw: true)
     end
 
     # @param entries [Hash]
-    # @param options [Hash]
+    # @option raw [Boolean]
     # @return [void]
     #
     # @api private
     # @since 0.3.0
     def write_multi(entries, **options)
+      raw = options.fetch(:raw, false)
+      entries = transform_pairset(entries) unless raw
+
       mset(*entries.to_a.flatten!, raw: true)
     end
   end